+++ tests/variable_definitions/test_vardef.py
@@ -6,8 +6,6 @@ from unittest.mock import patch
 import pytest
 import ruamel.yaml
 
-from dapla_metadata.variable_definitions.utils.template import _get_workspace_dir
-
 yaml = ruamel.yaml.YAML()
 
 from dapla_metadata._shared.config import DAPLA_GROUP_CONTEXT
@@ -290,45 +288,3 @@ def test_write_template_path_no_env_value(tmp_path: Path):
     result_without_timestamp += result.split(".yaml", 1)[1]
     expected_result = f"File path {workspace_dir}/variable_definitions/variable_definition_template.yaml Successfully written to file"
     assert result_without_timestamp == expected_result
-
-
-@pytest.mark.usefixtures("set_temp_workspace")
-def test_write_existing_variable_to_file(
-    client_configuration: Configuration,
-):
-    VardefClient.set_config(client_configuration)
-    file_name = Vardef.write_variable_to_file(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-
-    target_path = _get_workspace_dir() / file_name
-
-    with target_path.open(encoding="utf-8") as f:
-        parsed_yaml = yaml.load(f)
-
-    assert parsed_yaml["variable_status"] == "DRAFT"
-    assert parsed_yaml["short_name"] == "landbak"
-
-
-@pytest.mark.usefixtures("set_temp_workspace")
-def test_shortname_and_id_in_filename(
-    client_configuration: Configuration,
-):
-    VardefClient.set_config(client_configuration)
-    file_name = Vardef.write_variable_to_file(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-
-    assert "variable_definition_landbak_wypvb3wd_" in str(file_name)
-
-
-@pytest.mark.usefixtures("set_temp_workspace")
-def test_shortname_and_id_not_in_filename(
-    client_configuration: Configuration,
-):
-    VardefClient.set_config(client_configuration)
-    file_name = Vardef.write_variable_to_file(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-
-    assert "variable_definition_" in str(file_name)
